Use AbstractNavigationStateRenderer to capture all navigation requests
