+++ src/main/java/com/vaadin/extension/VaadinObservabilityInstrumentationModule.java
@@ -2,8 +2,8 @@ package com.vaadin.extension;
 
 import com.google.auto.service.AutoService;
 
+import com.vaadin.extension.instrumentation.AfterNavigationStateRendererInstrumentation;
 import com.vaadin.extension.instrumentation.EventRpcHandlerInstrumentation;
-import com.vaadin.extension.instrumentation.JavascriptBootstrapUiInstrumentation;
 import com.vaadin.extension.instrumentation.UiInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
@@ -40,7 +40,7 @@ public class VaadinObservabilityInstrumentationModule extends InstrumentationMod
     public List<TypeInstrumentation> typeInstrumentations() {
         // TypeIntrumentation for this instrumentation module
         return asList(
-                new JavascriptBootstrapUiInstrumentation(),
+                new AfterNavigationStateRendererInstrumentation(),
                 new UiInstrumentation(),
                 new EventRpcHandlerInstrumentation()
         );
